Standardize the naming of business details components to match those of other page components
