+++ src/app/form/(formSteps)/business-details/3/BusinessDetailsStep3.test.tsx
@@ -1,4 +1,4 @@
-import BusinessDetails4 from "@/app/form/(formSteps)/business-details/4/BusinessDetails4";
+import BusinessDetailsStep3 from "@/app/form/(formSteps)/business-details/3/BusinessDetailsStep3";
 import { renderWithRouter } from "@/app/form/_utils/testUtils/renderWithRouter";
 import {
   testFillFromDataStore,
@@ -8,67 +8,68 @@ import {
 } from "@/app/form/_utils/testUtils/sharedTests";
 import { screen } from "@testing-library/react";
 
-const noHasBeenExcludedFromMedicaid: TestField = {
+const noHasUncollectedDebt: TestField = {
   name: "No",
-  dataStoreKey: "hasBeenExcludedFromMedicaid",
+  dataStoreKey: "hasUncollectedDebt",
   required: true,
   requiredErrorMessage: "This question is required",
   role: "radio",
   testValue: "false",
   expectedValue: "false",
   withinGroupName:
-    "Have you ever been excluded or suspended by OIG (Office of Inspector General) from participation in Medicare, Medicaid/NJ FamilyCare, or CHIP? Select one *",
+    "Do you have any uncollected debt to Medicare, Medicaid/NJ FamilyCare, or CHIP (Children's Health Insurance Program)? Select one *",
 };
 
-const yesHasBeenExcludedFromMedicaid: TestField = {
+const yesHasUncollectedDebt: TestField = {
   name: "Yes",
-  dataStoreKey: "hasBeenExcludedFromMedicaid",
+  dataStoreKey: "hasUncollectedDebt",
   required: true,
   requiredErrorMessage: "This question is required",
   role: "radio",
   testValue: "true",
   expectedValue: "true",
   withinGroupName:
-    "Have you ever been excluded or suspended by OIG (Office of Inspector General) from participation in Medicare, Medicaid/NJ FamilyCare, or CHIP? Select one *",
+    "Do you have any uncollected debt to Medicare, Medicaid/NJ FamilyCare, or CHIP (Children's Health Insurance Program)? Select one *",
 };
 
-const noHasBeenSuspendedFromMedicaid: TestField = {
+const noIsSubjectToPaymentSuspension: TestField = {
   name: "No",
-  dataStoreKey: "hasBeenSuspendedFromMedicaid",
+  dataStoreKey: "isSubjectToPaymentSuspension",
   required: true,
   requiredErrorMessage: "This question is required",
   role: "radio",
   testValue: "false",
   expectedValue: "false",
   withinGroupName:
-    "Have you ever had Medicare, Medicaid/NJ FamilyCare, or CHIP enrollment/participation suspended, denied, revoked, or terminated? Select one *",
+    "Have you ever been subject to a payment suspension under a federal health care program? Select one *",
 };
 
-const yesHasBeenSuspendedFromMedicaid: TestField = {
+const yesIsSubjectToPaymentSuspension: TestField = {
   name: "Yes",
-  dataStoreKey: "hasBeenSuspendedFromMedicaid",
+  dataStoreKey: "isSubjectToPaymentSuspension",
   required: true,
   requiredErrorMessage: "This question is required",
   role: "radio",
   testValue: "true",
   expectedValue: "true",
   withinGroupName:
-    "Have you ever had Medicare, Medicaid/NJ FamilyCare, or CHIP enrollment/participation suspended, denied, revoked, or terminated? Select one *",
+    "Have you ever been subject to a payment suspension under a federal health care program? Select one *",
 };
 
 const minimalTestFields: Array<TestField> = [
-  yesHasBeenExcludedFromMedicaid,
-  yesHasBeenSuspendedFromMedicaid,
+  yesHasUncollectedDebt,
+  yesIsSubjectToPaymentSuspension,
 ];
 const allTestFields: Array<TestField> = [
-  noHasBeenExcludedFromMedicaid,
-  noHasBeenSuspendedFromMedicaid,
-  yesHasBeenExcludedFromMedicaid,
-  yesHasBeenSuspendedFromMedicaid,
+  noHasUncollectedDebt,
+  noIsSubjectToPaymentSuspension,
+  yesHasUncollectedDebt,
+  yesIsSubjectToPaymentSuspension,
 ];
 
-describe("<BusinessDetails4 />", () => {
-  const renderFunction = () => renderWithRouter(<BusinessDetails4 />, "/form/business-details/4");
+describe("<BusinessDetailsStep3 />", () => {
+  const renderFunction = () =>
+    renderWithRouter(<BusinessDetailsStep3 />, "/form/business-details/3");
 
   it("saves fields to the data store on submit", async () => {
     await testSaveFieldsToDataStore(minimalTestFields, minimalTestFields, renderFunction, screen);
